refactor(plyphon): consolidate egui impls into a ui module
Move every egui-flavoured item into a dedicated `ui` module (ui::node
for the NodeUi impls, ui::param for the shared inspector rows,
ui::ref_ext for DspRefExtUi, ui::settings for DspSettingsTab), so the
`egui` feature holds with a single cfg gate on the module declaration
instead of attributes scattered through the node, param and ref_ext
modules. The rest of the crate is headless by construction, so new UI
code cannot accidentally leak egui into a --no-default-features build.

The relocated NodeUi impls reach node fields through their existing
public accessors (set_count, set_size, set_rate, set_freq_lag,
set_gain_lag), since field privacy no longer spans the module boundary.
Crate-root re-exports (DspRefExtUi, DspSettingsTab) are unchanged, so
consumers are unaffected.
