+++ crates/gantz_plyphon/src/node/bus.rs
@@ -2,8 +2,6 @@
 
 use gantz_ca::CaHash;
 use gantz_core::node::{ExprCtx, ExprResult, MetaCtx};
-#[cfg(feature = "egui")]
-use gantz_egui::{NodeCtx, NodeUi, NodeUiResponse, Registry, SocketDoc, SocketKind};
 use gantz_nodetag::NodeTag;
 use serde::{Deserialize, Serialize};
 
@@ -70,34 +68,5 @@ impl NodeDsp for Bus {
 impl ToNodeDsp for Bus {
     fn to_node_dsp(&self) -> Option<&dyn NodeDsp> {
         Some(self)
-    }
-}
-
-#[cfg(feature = "egui")]
-impl NodeUi for Bus {
-    fn name(&self, _: &dyn Registry) -> &str {
-        "~bus"
-    }
-
-    fn description(&self) -> Option<&'static str> {
-        Some("Synthdef boundary: edits beyond it leave this side's synth running")
-    }
-
-    fn ui(&mut self, _ctx: NodeCtx, uictx: egui_graph::NodeCtx) -> NodeUiResponse {
-        let framed =
-            uictx.framed(|ui, _sockets| ui.add(egui::Label::new("~bus").selectable(false)));
-        NodeUiResponse::new(framed)
-    }
-
-    fn socket_doc(&self, _: &dyn Registry, kind: SocketKind, _ix: usize) -> Option<SocketDoc> {
-        match kind {
-            SocketKind::Input => Some(SocketDoc::ty("signal").with_description(
-                "signal to carry across the synthdef boundary (any channel width)",
-            )),
-            SocketKind::Output => Some(
-                SocketDoc::ty("signal")
-                    .with_description("the same signal, entering the downstream synthdef"),
-            ),
-        }
     }
 }
