+++ crates/gantz_plyphon/src/node/mod.rs
@@ -1,9 +1,10 @@
 //! The DSP node types.
 //!
 //! Each node implements [`gantz_core::Node`] (a placeholder, since DSP nodes are
-//! inert in the Steel world), [`NodeDsp`](crate::NodeDsp) (the audio behaviour),
-//! [`ToNodeDsp`](crate::ToNodeDsp) (discovery), and `gantz_egui::NodeUi` (the
-//! GUI, `egui` feature). Their `~` keyword-name prefix marks them as dsp nodes.
+//! inert in the Steel world), [`NodeDsp`](crate::NodeDsp) (the audio behaviour)
+//! and [`ToNodeDsp`](crate::ToNodeDsp) (discovery). Their `gantz_egui::NodeUi`
+//! impls live in `crate::ui` (`egui` feature). The `~` keyword-name prefix
+//! marks them as dsp nodes.
 
 pub use bus::Bus;
 pub use lag::Lag;
