+++ crates/gantz_plyphon/src/node/out.rs
@@ -4,10 +4,6 @@ use std::hash::{Hash, Hasher};
 
 use gantz_ca::CaHash;
 use gantz_core::node::{ExprCtx, ExprResult, MetaCtx, RegCtx};
-#[cfg(feature = "egui")]
-use gantz_egui::{
-    InspectorRowsResponse, NodeCtx, NodeUi, NodeUiResponse, Registry, SocketDoc, SocketKind,
-};
 use gantz_nodetag::NodeTag;
 use plyphon::Rate;
 use plyphon::synthdef::{InputRef, UnitSpec};
@@ -15,8 +11,6 @@ use serde::{Deserialize, Serialize};
 
 use crate::dsp::{DspBuilder, NodeDsp, Signal, ToNodeDsp};
 use crate::param::{cahash_lag, control_input_expr, param_name, param_state, plyphon_param};
-#[cfg(feature = "egui")]
-use crate::param::{param_row, param_state_row, param_value, with_value};
 
 /// The audio output sink. Applies a master `gain` to its input signal and writes
 /// it to the output buses. A mono input is fanned across every device channel; a
@@ -180,72 +174,6 @@ impl NodeDsp for Out {
 impl ToNodeDsp for Out {
     fn to_node_dsp(&self) -> Option<&dyn NodeDsp> {
         Some(self)
-    }
-}
-
-#[cfg(feature = "egui")]
-impl NodeUi for Out {
-    fn name(&self, _: &dyn Registry) -> &str {
-        "~out"
-    }
-
-    fn description(&self) -> Option<&'static str> {
-        Some("Audio output: master gain to the speakers")
-    }
-
-    fn ui(&mut self, _ctx: NodeCtx, uictx: egui_graph::NodeCtx) -> NodeUiResponse {
-        // The body shows just the node name; params are edited in the inspector.
-        let framed =
-            uictx.framed(|ui, _sockets| ui.add(egui::Label::new("~out").selectable(false)));
-        NodeUiResponse::new(framed)
-    }
-
-    fn show_state(&self) -> bool {
-        // A summarised "N queued" state row (in `inspector_rows`) replaces the raw
-        // `{value, pending}` dump.
-        false
-    }
-
-    fn inspector_rows(
-        &mut self,
-        ctx: &mut NodeCtx,
-        body: &mut egui_extras::TableBody,
-    ) -> InspectorRowsResponse {
-        let mut resp = InspectorRowsResponse::default();
-        let state = ctx.extract_value().ok().flatten();
-        param_state_row(body, state.as_ref());
-        let mut value = state
-            .as_ref()
-            .and_then(param_value)
-            .unwrap_or(Self::DEFAULT_GAIN as f64) as f32;
-        let mut lag = self.gain_lag;
-        let dv = egui::DragValue::new(&mut value)
-            .range(0.0..=1.0)
-            .speed(0.005);
-        let (value_changed, lag_changed) = param_row(body, "gain", dv, &mut lag);
-        if value_changed {
-            // Preserve any queued `pending` updates; only the value changes.
-            let prev = state.unwrap_or_else(|| param_state(Self::DEFAULT_GAIN as f64));
-            let _ = ctx.update_value(with_value(prev, value as f64));
-        }
-        if lag_changed {
-            self.gain_lag = lag;
-            resp.mark_changed();
-        }
-        resp
-    }
-
-    fn socket_doc(&self, _: &dyn Registry, kind: SocketKind, ix: usize) -> Option<SocketDoc> {
-        match (kind, ix) {
-            (SocketKind::Input, 0) => Some(SocketDoc::ty("signal").with_description(
-                "signal to send to the audio output; mono fans across every \
-                device channel, wider signals write channel i to bus i",
-            )),
-            (SocketKind::Input, 1) => Some(SocketDoc::ty("number").with_description(
-                "master gain control; overrides the inspector value while connected",
-            )),
-            _ => None,
-        }
     }
 }
